+++ NEWS.md
@@ -11,6 +11,7 @@ Bug fixes:
 + Properly infer the literal string value of `__FUNCTION__` for global functions in namespaces (#4231)
 + Fix false positive `PhanPossiblyInfiniteLoop` for `do {} while (false);` that is unchangeably false (#4236)
 + Infer that array_shift and array_pop return null when the passed in array could be empty, not false. (#4239)
++ Handle `PhpToken::getAll()` getting renamed to `PhpToken::tokenize()` in PHP 8.0.0RC4. (#4189)
 
 Oct 12 2020, Phan 3.2.3
 -----------------------